:recycle: Moved Header and UnverifiedToken into own module
